backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.3.0
fastapi==0.117.1
h11==0.16.0
idna==3.10
motor==3.7.1
pydantic==2.11.9
pydantic_core==2.33.2
pymongo==4.15.1
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
aiofiles==24.1.0
python-multipart==0.0.20
anthropic==0.34.0
google-cloud-vision==3.7.2
httpx==0.27.0
Pillow==10.4.0
reportlab==4.2.2
